Refactor content scripts, import & exports
Move myTeam.js and classicLeague.js to file contentScript.js and
create fpl.js that exports functions needed in contentScript.js.
