+++ src/manifest.json
@@ -17,14 +17,8 @@
   },
     "content_scripts": [
       {
-        "matches": ["*://fantasy.premierleague.com/a/team/my*"],
-        "js": ["myTeam.bundle.js"],
-        "run_at": "document_idle"
-      },
-      {
-        "matches": ["https://fantasy.premierleague.com/a/leagues/standings/*/classic",
-          "https://fantasy.premierleague.com/a/leagues/standings/*/classic*"],
-        "js": ["classicLeague.bundle.js"],
+        "matches": ["*://fantasy.premierleague.com/*"],
+        "js": ["contentScript.bundle.js"],
         "run_at": "document_idle"
       }
     ],
