+++ webpack.config.js
@@ -7,8 +7,7 @@ const path = require('path');
 module.exports = {
   entry: {
     background: path.resolve(__dirname, 'src/js/background.js'),
-    myTeam: path.resolve(__dirname, 'src/js/myTeam.js'),
-    classicLeague: path.resolve(__dirname, 'src/js/classicLeague.js'),
+    contentScript: path.resolve(__dirname, 'src/js/contentScript.js'),
     options: path.resolve(__dirname, 'src/js/options.js'),
     popup: path.resolve(__dirname, 'src/js/popup.js'),
   },
